+++ Common/Integrations/GenericModConfigMenu/IGenericModConfigMenuApi.cs
@@ -14,7 +14,7 @@ namespace Pathoschild.Stardew.Common.Integrations.GenericModConfigMenu
         void RegisterSimpleOption(IManifest mod, string optionName, string optionDesc, Func<string> optionGet, Action<string> optionSet);
         void RegisterChoiceOption(IManifest mod, string optionName, string optionDesc, Func<string> optionGet, Action<string> optionSet, string[] choices);
         void RegisterClampedOption(IManifest mod, string optionName, string optionDesc, Func<int> optionGet, Action<int> optionSet, int min, int max);
-        void RegisterClampedOption(IManifest mod, string optionName, string optionDesc, Func<float> optionGet, Action<float> optionSet, float min, float max);
+        void RegisterClampedOption(IManifest mod, string optionName, string optionDesc, Func<float> optionGet, Action<float> optionSet, float min, float max, float interval);
         void SetDefaultIngameOptinValue(IManifest mod, bool optedIn);
     }
 }